sync diffs

+++ data/data_processing/extract_images.py
@@ -10,6 +10,7 @@ import contextily as ctx
 from shapely.geometry import Point
 
 
+
 #image_path_nor = "/cluster/home/terjenf/MapTR/NAP_data/nuscenes/samples/C1_front60Single/frame_0782.png"
 trip_path = "/cluster/home/terjenf/MapTR/NAP_raw_data/Trip077/"
 root_folder = "/cluster/home/terjenf/MapTR/NAP_raw_data/"
@@ -17,6 +18,8 @@ root_folder = "/cluster/home/terjenf/MapTR/NAP_raw_data/"
 
 save_img_rooth = "/cluster/home/terjenf/NAPLab_car/data_images"
 folder_name ="Trip077"
+
+image__frame_span = (12672, 15972)
 #camera_json = "/cluster/home/terjenf/MapTR/NAP_raw_data/Trip077/camerasandCanandGnssCalibratedAll_lidars00-virtual.json"
 
 import utils
@@ -35,8 +38,6 @@ def create_folders(folder_name, cam_names):
     return new_folders 
 
 
-
-
 def extract_images(folder_name, camera_files, timestamp_files):
     """ 
     Extract images from files. Save image name with corresponding timestamp
@@ -50,14 +51,13 @@ def extract_images(folder_name, camera_files, timestamp_files):
         cam_folder_paths = [os.path.join(save_img_rooth,folder_name, cam_name) for cam_name in cam_names]
 
 
-    for camera_file, timestamp_file, cam_folder_path in zip(cmaera_files, timestamp_files, cam_folder_paths): 
+    for camera_file, timestamp_file, cam_folder_path, cam in zip(camera_files, timestamp_files, cam_folder_paths, cam_names): 
 
         count, time_stamps_cam = utils.get_timestamps(timestamp_file)
 
-        save_images_from_camera(cam_folder_path,camera_file, time_stamps_cam )
-
-        
-def save_images_from_camera(cam_folder_path, camera_file, time_stamps_cam):
+        save_images_from_camera(cam, cam_folder_path,camera_file, time_stamps_cam )
+     
+def save_images_from_camera(cam, cam_folder_path, camera_file, time_stamps_cam):
 
     """
     Extraxt and save images.
@@ -69,10 +69,8 @@ def save_images_from_camera(cam_folder_path, camera_file, time_stamps_cam):
  
     """
 
-
-
     cam_cap = cv2.VideoCapture(camera_file)
-    print("FPS: ", cap_2.get(cv2.CAP_PROP_FPS))
+    print("FPS: ", cam_cap.get(cv2.CAP_PROP_FPS))
 
 
     if not cam_cap.isOpened():
@@ -80,12 +78,19 @@ def save_images_from_camera(cam_folder_path, camera_file, time_stamps_cam):
     else:
         frame_count = 0
         while True:
+
+            if  not (image__frame_span[0] < frame_count < image__frame_span[1]):
+                continue
             # Read each frame
             ret, frame = cam_cap.read()
             if not ret:
                 break
             # Save the frame as an image file
-            frame_filename = os.path.join(cam_folder_path, f"frame_{frame_count:04d}_{time_stamps_cam[frame_count]}.png")
+            frame_filename = os.path.join(cam_folder_path, f"frame_{frame_count:04d}_{cam}_{time_stamps_cam[frame_count]}.png")
+
+            if os.path.exists(frame_filename): 
+                continue
+
             cv2.imwrite(frame_filename, frame)
             print(f"Saved: {frame_filename}")
 
@@ -104,15 +109,15 @@ if __name__ == "__main__":
 
     timestamps_files = utils.get_files(absoulute_files, file_format="timestamps")
 
-    cam_names = [cam.split("/")[-1].split(".")[0] for cam in camera_files]
+    #cam_names = [cam.split("/")[-1].split(".")[0] for cam in camera_files]
 
-
-
-
-    create_folders(folder_name, cam_names)
+    # create_folders(folder_name, cam_names)
 
 
     count, time_stamps_cam = utils.get_timestamps(timestamps_files[0])
+
+
+    extract_images(folder_name, camera_files, timestamps_files)
